# docs/openapi/api.v1.json
{
  "components" : {
    "parameters" : {
      "CorrelationId" : {
        "description" : "Optional trace ID; generated if absent and returned on every response",
        "in" : "header",
        "name" : "X-Correlation-Id",
        "required" : false,
        "schema" : {
          "example" : "demo-123",
          "type" : "string"
        }
      }
    },
    "schemas" : {
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "correlationId" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "violations" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            },
            "type" : "array"
          }
        }
      },
...
      "NoteRequest" : {
        "properties" : {
          "body" : {
            "maxLength" : 4000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "body" ],
        "type" : "object"
      },
...
      "PageResponseStaffClaimSummary" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/StaffClaimSummary"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ReasonRequest" : {
        "properties" : {
          "reason" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "reason" ],
        "type" : "object"
      },
      "ReassignRequest" : {
        "properties" : {
          "adjusterId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reason" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "adjusterId" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/v1/auth/login (valid 15 minutes)",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Guidewire-style claims handling: FNOL, exposures, reserves, payments, approvals, SIU and AI document assessment.\n\nAPI version **v1** (contract 1.3.0). Breaking changes get `/api/v2`; additions bump the minor version.",
    "title" : "AI Claims Platform API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/claims" : {
      "get" : {
        "description" : "Adjusters always get their own claims (assigneeId is ignored); supervisors see all.",
        "operationId" : "listClaims",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "SUBMITTED", "ASSESSING", "OPEN", "AWAITING_INFO", "SIU_REVIEW", "CLOSED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "segment",
          "required" : false,
          "schema" : {
            "enum" : [ "FAST_TRACK", "STANDARD", "COMPLEX" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "assigneeId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseStaffClaimSummary"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Work queue, newest first",
        "tags" : [ "Claims" ]
      },
      "post" : {
        "description" : "contactName is required. Idempotent with Idempotency-Key, as in the portal.",
        "operationId" : "reportLossByPhone",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FnolRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffClaimResponse"
                }
              }
            },
            "description" : "Claim created (or replayed); Location and ETag headers set"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Well-formed, but breaks a business rule"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Report a loss taken by phone (FNOL)",
        "tags" : [ "Claims" ]
      }
    },
...
    "/api/v1/claims/{id}" : {
      "get" : {
        "description" : "ETag = version; send it as If-Match on commands",
        "operationId" : "getClaim",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffClaimResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Claim detail",
        "tags" : [ "Claims" ]
      }
    },
    "/api/v1/claims/{id}/close" : {
      "post" : {
        "description" : "OPEN -> CLOSED (assigned adjuster)",
        "operationId" : "closeClaim",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffClaimResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "412" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "`If-Match` does not match the current version: reload and retry"
          },
          "428" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "`If-Match` is required on this request: send the ETag from your last GET"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Close the claim",
        "tags" : [ "Claims" ]
      }
    },
    "/api/v1/claims/{id}/notes" : {
      "post" : {
        "operationId" : "addClaimNote",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NoteResponse"
                }
              }
            },
            "description" : "Note added"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Add an internal note (never shown to the claimant)",
        "tags" : [ "Claims" ]
      }
    },
    "/api/v1/claims/{id}/reassign" : {
      "post" : {
        "operationId" : "reassignClaim",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReassignRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffClaimResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "412" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "`If-Match` does not match the current version: reload and retry"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Well-formed, but breaks a business rule"
          },
          "428" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "`If-Match` is required on this request: send the ETag from your last GET"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Assign to another adjuster (supervisor)",
        "tags" : [ "Claims" ]
      }
    },
    "/api/v1/claims/{id}/reopen" : {
      "post" : {
        "description" : "CLOSED -> OPEN (supervisor)",
        "operationId" : "reopenClaim",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffClaimResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "412" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "`If-Match` does not match the current version: reload and retry"
          },
          "428" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "`If-Match` is required on this request: send the ETag from your last GET"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Reopen a closed claim",
        "tags" : [ "Claims" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8081"
  } ],
...
}